Move spatialdata helper functions to new file
